Bump the `moment` dependency from v2.29.3 to v2.29.4
This updates the `moment` dependency from v2.29.3 to v2.29.4.

This version update addresses a [CVE][1] published in the
GitHub Advisory Database, where a regular expression used in
moment can leave the door open to DDoS attacks. This is
unlikely to affect the Obsidian use case, but using this
vulnerable version triggers security warnings (for example
from GitHub's Dependabot) and the upgrade should be
harmless given that the DDoS fix is the only thing in the
[changelog][2] for this version.

[1]: https://github.com/advisories/GHSA-wc69-rhjr-hc9g