More descriptive canvas type definitions
The main change here is to convert the comments to JSDoc. This presents
no changes when reading the file, but when using an IDE to look at the
types we're presented with the comments as supplementary info in a
tooltip etc. Very useful!

As well as this, we pull in the Join type from ts-toolbelt, which allows
us to enforce the type of `CanvasFileData["subpath"]` so it complies
with the description of it in the comment.

Lastly, we derive some types in `CanvasEdgeData` instead of using
`string`, which encodes the intent as well as the type and makes it less
likely to break in the future.